// src/types/modules/categoryStatsVO.ts
/**
 * // Code generated by orval. DO NOT EDIT.
 * // 服务: Openapi在线JSON文档
 * // 版本: 1.0.0
 * // 生成时间: 2026-09-21T17:12:14.856Z
 */

/**
 * 分类统计
 */
export interface CategoryStatsVO {
  /** 分类ID */
  id?: number;
  /** 分类名称 */
  name?: string;
  /** 分类描述 */
  desc?: string;
  /** 分类类型: 0-导航栏分类, 1-普通分类 */
  type?: number;
  /** 排序值 */
  sort?: number;
  /** 图标 */
  icon?: string;
  /** 该分类下的文章数量 */
  countOfSort?: number;
}