// src/types/modules/categoryWebVO.ts
/**
 * // Code generated by orval. DO NOT EDIT.
 * // 服务: Openapi在线JSON文档
 * // 版本: 1.0.0
 * // 生成时间: 2026-10-05T08:15:14.299Z
 */

/**
 * 分类公开VO
 */
export interface CategoryWebVO {
  /** ID */
  id?: number;
  /** 分类名称 */
  name?: string;
  /** 分类描述 */
  desc?: string;
  /** 排序 */
  sort?: number;
  /** 图标 */
  icon?: string;
}